Fatal error: need to convert size_t to int before going into a minus operator

+++ src/Tracker.cxx
@@ -231,7 +231,7 @@ int Tracker::tracking(int iLayer,size_t & iselection){
     else{
         for (int i = -1; i<(int)hitLayerIndexMap->at(iLayer)->size(); i++){ // -1 means don't pick up any hit in this layer 
             if (i==-1){
-                if (hitLayerIndexMap->at(iLayer)->size()==0||(skipLayerAllowed&&NLAY-iLayer-1+currentTrack3D.hitIndexSelected.size()-currentTrack3D.nPars()>=1)){
+                if (hitLayerIndexMap->at(iLayer)->size()==0||(skipLayerAllowed&&NLAY-iLayer-1+(int)currentTrack3D.hitIndexSelected.size()-currentTrack3D.nPars()>=1)){
                     MyNamedVerbose("Tracking",Form(" => skip layer %d",iLayer));
                     tracking(iLayer+1,iselection);
                 }
